More warning fixes

+++ lib/Xmu/StrToCurs.c
@@ -50,7 +50,7 @@ SOFTWARE.
 
 ******************************************************************/
 
-/* $XFree86: xc/lib/Xmu/StrToCurs.c,v 1.9tsi Exp $ */
+/* $XFree86: xc/lib/Xmu/StrToCurs.c,v 1.10tsi Exp $ */
 
 #include	<X11/Intrinsic.h>
 #include	<X11/StringDefs.h>
@@ -168,8 +168,9 @@ XmuCvtStringToCursor(XrmValuePtr args, Cardinal *num_args,
 #endif
 
 	strspec = XtMalloc(strlen("FONT %s %d %s %d") + 21);
-	sprintf(strspec, "FONT %%%lds %%d %%%lds %%d", sizeof(source_name) - 1,
-		sizeof(mask_name) - 1);
+	sprintf(strspec, "FONT %%%lds %%d %%%lds %%d",
+		(ulong)sizeof(source_name) - 1,
+		(ulong)sizeof(mask_name) - 1);
 	fields = sscanf(name, strspec,
 			source_name, &source_char,
 			mask_name, &mask_char);
